requirements.txt
flask 
scikit-learn
pandas
numpy 
xgboost
catboost
psycopg2
dill

python-dotenv